# util/sha256: new package